Add files via upload
Coding Challenge solution for the first problem: Business Card OCR. Added three function getName, getPhoneNumber, and getEmail. all of these functions require a string that is the business card read as a plain text and returns a string to be stored or displayed. Added the readme for setup, and a requirements.txt that can be used for quick deployment in visual studios.